+++ pom.xml
@@ -11,6 +11,7 @@
     <groupId>com.ddu</groupId>
     <artifactId>goushushenpixitong</artifactId>
     <version>0.0.1-SNAPSHOT</version>
+    <packaging>jar</packaging>
     <name>goushushenpixitong</name>
     <description>Demo project for Spring Boot</description>
 
@@ -144,11 +145,16 @@
     </dependencies>
 
     <build>
+
         <plugins>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
+                <configuration>
+                    <executable>true</executable>
+                </configuration>
             </plugin>
+
 
             <plugin>
                 <groupId>org.mybatis.generator</groupId>
